test: remove `strict` array snapshot files, move content to `base` snapshots
