+++ test/snapshots/json-schema-draft-04/advanced-example/_expected.d.ts
@@ -5,7 +5,10 @@ declare namespace SomeSiteSomewhere {
     export interface EntrySchema {
         storage: EntrySchema.Definitions.DiskDevice | EntrySchema.Definitions.DiskUUID | EntrySchema.Definitions.Nfs | EntrySchema.Definitions.Tmpfs;
         fstype?: "ext3" | "ext4" | "btrfs";
-        options?: string[];
+        options?: [
+            string,
+            ...string[]
+        ];
         readonly?: boolean;
     }
     namespace EntrySchema {
