--- test/snapshots/json-schema-draft-04/openapi-v2.0/_config_strict.json
@@ -1,3 +1,0 @@
-{
-  "strictArraySize": true
-}
