--- test/snapshots/json-schema-draft-04/schema/_config_strict.json
@@ -1,3 +1,0 @@
-{
-  "strictArraySize": true
-}
