--- test/snapshots/json-schema-draft-07/advanced-example/_config_strict.json
@@ -1,3 +1,0 @@
-{
-  "strictArraySize": true
-}
