+++ test/snapshots/json-schema-draft-07/simple-example/_expected.d.ts
@@ -16,7 +16,10 @@ declare namespace JsonSchemaOrg {
         id: number;
         name: string;
         price: number;
-        tags?: string[];
+        tags?: [
+            string,
+            ...string[]
+        ];
         dimensions?: {
             length: number;
             width: number;
